Module 8: Basic CNN and Transfer learning
Objective: The primary goal of this assignment is to enable students to consolidate and articulate key concepts, methodologies, and tools pertinent to object detection in a succinct and accessible manner. This will be achieved through the creation of a basic cheat sheet which will serve as a quick reference guide for object detection tasks.